refactor: extract usePagefind hook and SearchResults from Search
Search.tsx mixed the Pagefind loader/debounce/search machinery, the
modal chrome, and the results markup in one 587-line file. The data
side now lives in usePagefind (including the close-reset, where a
cancelled in-flight search must force isFetching off) and the listbox
markup in SearchResults; Search keeps the modal, focus management, and
tabs. Pure movement — every aria attribute, id, and translation key is
byte-identical, and the DOM tests pass untouched.
